chanes to file structure & auto initialization
breaking up utilities module into a package with a same name.

some changes to main file.
now there is no need to run initialize script first , main script handle initialization if needed
