+++ core.py
@@ -1,9 +1,10 @@
 import gensim
-import utilities
+from utilities import tools
+from utilities import nlp
 import numpy as np
-import datetime
 from os import listdir
 from os.path import join, isfile
+
 
 class Tony():
     def __init__(self, model_addr, tfidf_addr, phrases_addr, knowledge_dir='knowledge', threshold=0.01, tfidf_factor=1):
@@ -12,16 +13,14 @@ class Tony():
         self.tfidf_factor = tfidf_factor
         self.knowledge = dict()
 
-
         self.model = gensim.models.Word2Vec.load(model_addr)
 
-        self.tfidf = utilities.tfidf()
+        self.tfidf = nlp.tfidf()
         self.tfidf.load()
 
         self.bigram_transformer = gensim.models.phrases.Phraser.load(phrases_addr)
 
         self.code_knowledge()
-
 
     def set_knowledge_dir(self, knowledge_dir):
         self.knowledge_dir = knowledge_dir
@@ -32,10 +31,12 @@ class Tony():
     def new_representation(self, questions):
         m = len(questions)
         n = self.model.vector_size
-        X = np.zeros((m,n))
+        X = np.zeros((m, n))
         for i, question in enumerate(questions):
-            X[i] = np.sum([self.model[word]/utilities.l2(self.model[word]) * (self.tfidf[word] if self.tfidf.get(word) else 1) ** self.tfidf_factor
-                           for word in self.bigram_transformer[question.split()] if word in self.model.wv.vocab], axis= 0)
+            X[i] = np.sum([self.model[word] / tools.l2(self.model[word]) * (
+                self.tfidf[word] if self.tfidf.get(word) else 1) ** self.tfidf_factor
+                           for word in self.bigram_transformer[question.split()] if word in self.model.wv.vocab],
+                          axis=0)
         return X
 
     def code_knowledge(self):
@@ -44,7 +45,6 @@ class Tony():
             with open(join(self.knowledge_dir, file_)) as f:
                 self.knowledge[file_] = self.new_representation(f.readlines())
 
-
     def answer(self, question):
         answers = list()
         summary = list()
@@ -52,11 +52,11 @@ class Tony():
 
         for doc_name, doc_matrix in self.knowledge.iteritems():
             for sentence_num, sentence in enumerate(doc_matrix):
-                sim = utilities.cosine_dist(question_vec, sentence.reshape(-1))
+                sim = tools.cosine_dist(question_vec, sentence.reshape(-1))
                 if sim > self.threshold:
                     answers.append((doc_name, sentence_num, sim))
 
-        sorted_answers = sorted(answers, key= lambda x: x[2], reverse=True)
+        sorted_answers = sorted(answers, key=lambda x: x[2], reverse=True)
 
         for (doc_name, sentence_num, _) in sorted_answers:
             with open(join(self.knowledge_dir, doc_name)) as f:
