+++ main.py
@@ -1,12 +1,19 @@
 import core
-import datetime
-
+import os
+import initialize
 
 while True:
+    if not os.path.exists("Phrases") and not os.path.exists("tfidf.pkl"):
+        print "intilizing"
+        initialize.start()
+
+    print "how can i help you ?"
     question = raw_input()
-    #before = datetime.datetime.now()
+
     tony = core.Tony('word2vec/400', 'tfidf', 'Phrases', tfidf_factor=0.5)
-    #print 'How can I help you?'
-    print tony.answer(question)[0]
-    #after = datetime.datetime.now()
-    #print after - before
+
+    answer = tony.answer(question)[0]
+    if not answer:
+        print answer
+    else:
+        print "i'm not sure , i will ask him !"
